Separated DAY_TICKS and VANILLA_DAY_TICKS. * Introduced VANILLA_DAY_TICKS * DAY_TICKS is now a macro and calls ::GetDayTicks() * Introduced GetPaceFactor() currently it returns 1. In future commits it will use the game creation settings value. * All trivial cases where we need VANILLA_DAY_TICKS also have been updated. * There are some mechs which we will update in future commits.
Also few notes:

* Cargo transit days uses vanilla days.
* Timetables is bound to animation pace and mech itself will remain without changes. timetable_gui though is a subject of separate commit.
* Network bound to animation pace, and uses VANILLA_DAY_TICKS.

+++ src/town.h
@@ -73,7 +73,7 @@ struct Town : TownPool::PoolItem<&_town_pool> {
 	int16 ratings[MAX_COMPANIES];  ///< ratings of each company for this town
 
 	TransportedCargoStat<uint32> supplied[NUM_CARGO]; ///< Cargo statistics about supplied cargo.
-	TransportedCargoStat<uint16> received[NUM_TE];    ///< Cargo statistics about received cargotypes.
+	TransportedCargoStat<uint32> received[NUM_TE];    ///< Cargo statistics about received cargotypes.
 	uint32 goal[NUM_TE];                              ///< Amount of cargo required for the town to grow.
 
 	std::string text; ///< General text with additional information.
@@ -82,10 +82,10 @@ struct Town : TownPool::PoolItem<&_town_pool> {
 
 	StationList stations_near;       ///< NOSAVE: List of nearby stations.
 
-	uint16 time_until_rebuild;       ///< time until we rebuild a house
+	uint32 time_until_rebuild;       ///< time until we rebuild a house
 
-	uint16 grow_counter;             ///< counter to count when to grow, value is smaller than or equal to growth_rate
-	uint16 growth_rate;              ///< town growth rate
+	uint32 grow_counter;             ///< counter to count when to grow, value is smaller than or equal to growth_rate
+	uint32 growth_rate;              ///< town growth rate
 
 	byte fund_buildings_months;      ///< fund buildings program in action?
 	byte road_build_months;          ///< fund road reconstruction in action?
